+++ demo/src/main/java/com/example/demo/servicio/ClienteService.java
@@ -6,13 +6,13 @@ import com.example.demo.entidad.Cliente;
 
 public interface ClienteService {
 
-    public Cliente SearchById(int id);
+    public Cliente SearchById(Long id);
 
     public Cliente SearchByCedula(String cedula);
 
     public Collection<Cliente> SearchAll();
 
-    public void deleteById(int id);
+    public void deleteById(Long id);
 
     public void update(Cliente cliente);
